requirements.txt
Flask
Flask-Mail
Werkzeug
flask_sqlalchemy
gunicorn

psycopg2-binary

python-dotenv